move old code to backup dir when upgrading agent
agent_wrap changes

agent_wrap changes

renamed agent_storage to noobaa_storage

rename agent_storage to noobaa_storage, and move to the root of the drive\mount

added messages

incrment version to 1.12.0

removed redundant code

in windows reduce the number of time waiting for script end

fixed indentation

fixed by review comments

fixed log message

+++ src/deploy/NVA_build/clean_ova.sh
@@ -61,7 +61,12 @@ sudo cp -f /root/node_modules/noobaa-core/src/deploy/NVA_build/noobaa_supervisor
 sudo cp -f /root/node_modules/noobaa-core/src/deploy/NVA_build/env.orig /root/node_modules/noobaa-core/.env
 supervisorctl reread
 supervisorctl reload
-rm -rf /root/node_modules/noobaa-core/agent_storage/
+if [ -d /root/node_modules/noobaa-core/agent_storage/ ]; then
+    rm -rf /root/node_modules/noobaa-core/agent_storage/
+fi
+if [ -d /root/node_modules/noobaa-core/noobaa_storage/ ]; then
+    rm -rf /root/node_modules/noobaa-core/noobaa_storage/
+fi
 sleep 15
 mongo nbcore --eval 'db.dropDatabase()'
